requirements.txt
google-api-python-client
serpapi
openpyxl
xlsxwriter
python-dotenv
fastapi
uvicorn
websockets
